backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --no-dev --format requirements-txt
-e .
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via strands-agents-tools
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
aws-requests-auth==0.4.3
    # via strands-agents-tools
beautifulsoup4==4.14.3
    # via markdownify
bedrock-agentcore==1.1.2
    # via booking-backend
boto3==1.42.17
    # via
    #   bedrock-agentcore
    #   booking-backend
    #   strands-agents
botocore==1.42.17
    # via
    #   bedrock-agentcore
    #   boto3
    #   s3transfer
    #   strands-agents
    #   strands-agents-tools
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   uvicorn
cryptography==46.0.3
    # via pyjwt
dill==0.4.0
    # via strands-agents-tools
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via strands-agents
email-validator==2.3.0
    # via booking-backend
fastapi==0.128.0
    # via booking-backend
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.1
    # via opentelemetry-api
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.25.1
    # via
    #   mcp
    #   strands-agents
jsonschema-specifications==2025.9.1
    # via jsonschema
markdown-it-py==4.0.0
    # via rich
markdownify==1.2.2
    # via strands-agents-tools
mcp==1.25.0
    # via strands-agents
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
opentelemetry-api==1.39.1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-threading
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   strands-agents
opentelemetry-instrumentation==0.60b1
    # via opentelemetry-instrumentation-threading
opentelemetry-instrumentation-threading==0.60b1
    # via strands-agents
opentelemetry-sdk==1.39.1
    # via strands-agents
opentelemetry-semantic-conventions==0.60b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==25.0
    # via opentelemetry-instrumentation
pillow==11.3.0
    # via strands-agents-tools
prompt-toolkit==3.0.52
    # via strands-agents-tools
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   bedrock-agentcore
    #   booking-backend
    #   fastapi
    #   mcp
    #   pydantic-settings
    #   strands-agents
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via mcp
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via
    #   booking-backend
    #   mcp
    #   strands-agents-tools
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.2.1
    # via
    #   booking-backend
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.21
    # via mcp
pywin32==311 ; sys_platform == 'win32'
    # via mcp
pyyaml==6.0.3
    # via uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   aws-requests-auth
    #   strands-agents-tools
rich==14.2.0
    # via strands-agents-tools
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via
    #   markdownify
    #   python-dateutil
slack-bolt==1.27.0
    # via strands-agents-tools
slack-sdk==3.39.0
    # via slack-bolt
soupsieve==2.8.1
    # via beautifulsoup4
sse-starlette==3.1.1
    # via mcp
starlette==0.50.0
    # via
    #   bedrock-agentcore
    #   fastapi
    #   mcp
    #   sse-starlette
strands-agents==1.20.0
    # via
    #   booking-backend
    #   strands-agents-tools
strands-agents-tools==0.2.18
    # via booking-backend
sympy==1.14.0
    # via strands-agents-tools
tenacity==9.1.2
    # via strands-agents-tools
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   bedrock-agentcore
    #   fastapi
    #   mcp
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   strands-agents
    #   strands-agents-tools
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2025.3 ; sys_platform == 'win32'
    # via strands-agents-tools
urllib3==2.6.2
    # via
    #   bedrock-agentcore
    #   botocore
    #   requests
uvicorn==0.40.0
    # via
    #   bedrock-agentcore
    #   booking-backend
    #   mcp
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchdog==6.0.0
    # via
    #   strands-agents
    #   strands-agents-tools
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.2.14
    # via prompt-toolkit
websockets==15.0.1
    # via
    #   bedrock-agentcore
    #   uvicorn
wrapt==1.17.3
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-threading
yarl==1.22.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
